requirements.txt
selenium
PyYAML
lazy-load